Separated validators into different files. Updated message templates api
