+++ src/routes/messageTemplate.api.ts
@@ -4,6 +4,7 @@ import auth from '../middleware/auth';
 import wrapAsync from '../utils/asyncWrapper';
 import { MessageTemplate } from '../models/messageTemplate.model';
 import { ValidationError } from '../exceptions';
+import { validateLanguage, validateMessageTemplateType, validateMessageTime, validateMongoId } from '../validators';
 
 const router = express.Router();
 
@@ -14,6 +15,8 @@ router.post('/newTemplate', auth, wrapAsync(async (req, res) => {
   if (!messageTxt) {
     new ValidationError('Please Enter Message Text!');
   }
+  validateLanguage(language);
+  validateMessageTemplateType(type);
 
   const newMessageTemplate = new MessageTemplate({
     language: language,
@@ -28,12 +31,16 @@ router.post('/newTemplate', auth, wrapAsync(async (req, res) => {
   });
 }));
 
-router.post('/deleteTemplate', wrapAsync(async (req, res) => {
+router.post('/deleteTemplate', auth, wrapAsync(async (req, res) => {
   const { id } = req.body;
 
-  await MessageTemplate.findByIdAndDelete(new ObjectId(id));
+  if(!validateMongoId(id)) {
+    throw new ValidationError("Invalid templateId");
+  }
 
-  res.status(200);
+  await MessageTemplate.findByIdAndDelete(id);
+
+  res.status(200).json();
 }));
 
 router.get('/templates', auth, wrapAsync(async (req, res) => {
